test(web): split oversized dockview action tests
